Split off setuptools adapter
Authored with Claude.
